修改成了单例,是析构函数导致的事件丢失导致画笔被中断

+++ Project/AiferuFramework/Assets/AiferuFramework/库本身的功能/4.项目使用工具收集/8.美术刷草工具/ArtBrushTool_Aiferu/Script/ArtBrushToolData.cs
@@ -25,7 +25,7 @@ namespace AiferuFramework.ArtBrushTool
         /// <summary>
         /// 画刷大小
         /// </summary>
-        public int BrushSize;
+        public float BrushSize;
         /// <summary>
         /// 草对象随机大小范围最大值
         /// </summary>
